Move feature gates to a separate pkg
Moving feature gates implemented in the apiserver to a dedicated package
can help in using it in different places. For example, in controller-manager.
